feat(redis): add a generic migrations system
This move out from the worker the migration system.

It changes how we stamps the fact we have done the migration script.
